+++ src/constants.ts
@@ -1,4 +1,5 @@
 export const ENV = process.env.NODE_ENV
 
 export const BACKEND_URL = process.env.VUE_APP_BACKEND_URL
+export const APP_NAME = "mitra-web"
 export const APP_VERSION = process.env.VUE_APP_VERSION